layout re-org, React-route and FE routes built
