--- Client/login and signup/signup-login.css
@@ -1,4 +1,0 @@
-body {
-    background-color: whitesmoke;
-}
-
