--- Client/login and signup/signup-login.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-        <!-- the title is what is displayed on tab header -->
-        <title>Log In/Sign Up</title>
-        <link rel="stylesheet" href="./signup-login.css">
-    </head>
-    
-    <body>
-    <script>
-        function printMe(e) {
-            fileContents = document.getElementById("pw").value
-            // e.preventDefault()
-            // document.getElementById("my-div").style.backgroundColor = "whtesmoke";
-            console.log(fileContents);
-        }
-    </script>
-    
-    
-        <!-- website header and nav bar -->
-        <header>
-            <h1>Ready to get your stretch on?</h1>
-        </header>
-        <p>Don't have an account?</p>
-        <h2>Sign up</h2>
-        <!-- signup form -->
-        <form id="signup">
-            <label for="name">Name:</label>
-            <input type="text" id="name" name="name"><br><br>
-            <label for="email">Email address:</label>
-            <input type="text" id="lname" name="lname" value="Doe"><br><br>
-            <label for="password">Password:</label>
-            <input type="text" id="signup-pw" name="pw"><br><br>
-            <button onClick="printMe()">Sign up</button>
-        </form> 
-        <br>
-        <hr>
-        <p>Already have an account?</p>
-        <h2>Log in</h2>
-        <!-- login form -->
-        <form id="login">
-            <label for="email">Email address:</label>
-            <input type="text" id="email"><br><br>
-            <label for="password">Password:</label>
-            <input type="text" id="login-pw" name="pw"><br><br>
-            <button onClick="printMe()">Log in</button>
-        </form> 
-    </body>
-</html>
