+++ Client/styles.css
@@ -11,7 +11,7 @@ body {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-  overflow: hidden;
+  /* overflow: hidden; */
 }
 
 /*-------------------- HEADERCONTAINER STYLES --------------------*/
