+++ package.json
@@ -13,6 +13,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.4.0",
+    "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.6",
     "dotenv": "^16.3.1",
     "express": "^4.16.3",
     "mongodb": "^5.6.0",
